requirements.txt
telethon>=1.25.0
Pillow>=10.0.0
